Update script.js
Altering logic for multi-page patterns. 

Establishing the "default" as a 6 inch hoop with 14 count fabric. which works out to 84 stitches. Anything 84 stitches or less should print on a single page.

For multi-page prints, established a fixed scale based on the "most full" page, so that in theory the full pattern can be printed, cut out, and taped together.